项目开发plus/backend/pyproject.toml
[project]
name = "ai-express-workbench"
version = "0.1.0"
description = "多 Agent 协同开发可视化工作台"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sse-starlette>=2.1.0",
    "langgraph>=0.4.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0",
    "watchdog>=4.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aiexpress", "app"]
